fix: getAmountValue INT -> DOUBLE로 수정

+++ back/src/main/java/msps/back/dto/response/AllMenuIngredientsGetResponse.java
@@ -39,7 +39,7 @@ public class AllMenuIngredientsGetResponse {
         private String name;
         private String type;
         private AmountType amountType;
-        private Integer amountValue;
+        private Double amountValue;
         private String amountUnit;
         private String amountText;
     }
